+++ src/rotp/model/ships/ShipWeaponMissile.java
@@ -1,6 +1,6 @@
 /*
  * Copyright 2015-2020 Ray Fowler
- * Modifications Copyright 2023 Ilya Zushinskiy
+ * Modifications Copyright 2023-2025 Ilya Zushinskiy
  * 
  * Licensed under the GNU General Public License, Version 3 (the "License");
  * you may not use this file except in compliance with the License.
@@ -44,7 +44,7 @@ public final class ShipWeaponMissile extends ShipWeaponMissileType {
     @Override
     public TechMissileWeapon tech()       { return (TechMissileWeapon) super.tech(); }
     @Override
-    public float computerLevel()         { return tech().computer; }
+    public int computerLevel()            { return tech().computer; }
     @Override
     public Image image(int num)           { return tech().image(num); }
     @Override
